* commit 'ee04ec5455ea91bc1c059cddfe9c1e19c6a1b10e':
  orientations are correct for neighbors that cross periodic boundary, readdyFileInfo object is now passed through to calculateOrientations() in order to have access to boxSize while calculating
  fixed compile error in connection manager by adding cast to int

+++ src/main/connection_manager.cpp
@@ -541,8 +541,8 @@ namespace agentsim {
 
         Json::Value dataFrameMessage;
         dataFrameMessage["msgType"] = id_vis_data_arrive;
-        dataFrameMessage["bundleSize"] = numberOfFrames;
-        dataFrameMessage["bundleStart"] = startingFrame;
+        dataFrameMessage["bundleSize"] = (int)numberOfFrames;
+        dataFrameMessage["bundleStart"] = (int)startingFrame;
         Json::Value frameArr = Json::Value(Json::arrayValue);
         int endFrame = std::min(startingFrame + numberOfFrames, totalNumberOfFrames - 1);
